Initial Commit
This lays out the start of what is roughly in my head and what
I have been tinkering on for an ergonomic Elasticsearch Rust
client and query generation.  It insulates the direct use of
the offical client in favor of a more dumbed down set of options
that make the most sence to me so far.