Add Initial Support for Msearch
This implements multi-search in the most ergonomic way that I
would expect to see it done; which is being able to provide an
iteratorable structure that will correctly serialize into the
msearch request under the hood.

Something else I've done as a side feature is the structure
responsible for converting these can actually work with different
search types at the same time if the collection is made ahead
of time.

I feel this is a good first pass; however, another feature
addition pass is for sure in the works.  Items to solve in
future work include:

- Handling of a search that fails in in isolation
- Adding more ergonomic features to the MultiSearch structure