Introduce Breaking Ergonomic Filter Changes
This removes the long complicated chaining sequence that
was going to ultimately be very difficult to make composable.
Instead of the long chain with elects to have `with` accept
anything that can be converted into a `SearchCondition`.

All `Criterion` and by extension their builders fall into this
category after this change.  This has drastically lowered the
complexity of how criteria are placed into buckets and now
conditions can be generated by another process and fed into a
search as data instead of needing to do function transforms.

Ultimately I think this is a big win for the ergonomics.