Add Script-Score Sort Support & Refactor Sort
This originally started as adding in the feature for script-score
sorting. However; I have also added the same ergonomic changes to
the sort system that were made with adding criteria to a search.

Now instead of longer sort chains from the search it accepts any
builder into a `Sort`.  This allows them to be composable in the
same way criteria are.

The introduction of the script-score into this code base is not
in the final place I would like it to be.  The amount of logic
which happens to shift the search body for this special search is
quite a bit and needs a good refactor at some point.

While there are other items that could have added to the script-score
I have elected to only add in the minimum needed to get this feature
working.