Add `MultiResponse::to_inner(&self)`
Upon working with the multi-response it has become quite apparent
that there are ownership needs for the underlying data that are
needed.  Instead of trying to mimic all of the features of the
vector that contains the data I have elected to simply allow users
to unwrap the `MultiResponse` down to the actual data with `#to_inner`.