Fix `empty bool` Malformed Query Bug
Empty `bool` query building bug is fixed.  When attempting to build
a search if you apply an empty criteria set, such as an `AllMatch`
it would generate a malformed query where the `bool` node had an
empty filter.  Now criterion are checked before being applied to a
search to prevent this going forward.