Add `unwrap_as` and `borrow_as` to `AggResult` (#38)
Ergonomic methods to `AggResult` that allow you to directly get at
the data the enum is wrapping.  Prior to this change you had to work
directly from the `AggResultCollection`; which may not always be the
case if you're attempting to pass individual results around to be
processedd by the the code that created the request for them w/o
passing the entire collection.