🐛 Fix Aggregation Collection Deserialization
There are some keys which act as supplemental information
which would not be used currently by `elastic_lens`.  This
namely shows up when fetching terms for a `bool`.  Elasticsearch
provides a `key_as_string` key which holds a string representation
of the JSON value of the term.

We do not need these as we are already in a typed system so
it's safe to skip any keys which do not contain a collection
hint.