Add Script Sort Feature (#52)
Prior to this I forgot that ES has two different kinds of scripts
that you can provide to sort results.  One is specific for search
sore sorting and the other is just a means by which to generate
a scalar value to be used when sorting.  Until now only the search
score scripting was available.  This adds in support for both
kinds of sorts; however, it is a breaking change because the old
`by_script` is now performing a standard script score and not
the what it's new neighbor `by_script_score` is now doing.